pyproject.toml
[project]
name = "axa-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "contextily>=1.7.0",
    "duckdb>=1.4.3",
    "folium>=0.20.0",
    "geopandas>=1.1.2",
    "h3>=4.4.1",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "nbformat>=5.10.4",
    "networkx>=3.4.2",
    "numpy>=2.2.6",
    "osmnx>=2.0.7",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "polars>=1.36.1",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "shapely>=2.1.2",
    "statsmodels>=0.14.6",
]

[tool.setuptools.packages.find]
where = ["src"]
